Move writeScriptsTo to plutus-contract
