+++ nix/pkgs/haskell/materialized-windows/.plan.nix/plutus-contract.nix
@@ -54,7 +54,10 @@
           (hsPkgs."cryptonite" or (errorHandler.buildDepError "cryptonite"))
           (hsPkgs."data-default" or (errorHandler.buildDepError "data-default"))
           (hsPkgs."deepseq" or (errorHandler.buildDepError "deepseq"))
+          (hsPkgs."directory" or (errorHandler.buildDepError "directory"))
+          (hsPkgs."filepath" or (errorHandler.buildDepError "filepath"))
           (hsPkgs."fingertree" or (errorHandler.buildDepError "fingertree"))
+          (hsPkgs."flat" or (errorHandler.buildDepError "flat"))
           (hsPkgs."foldl" or (errorHandler.buildDepError "foldl"))
           (hsPkgs."freer-simple" or (errorHandler.buildDepError "freer-simple"))
           (hsPkgs."hashable" or (errorHandler.buildDepError "hashable"))
@@ -142,6 +145,7 @@
           "Plutus/Trace/Effects/Waiting"
           "Plutus/Trace/Emulator"
           "Plutus/Trace/Emulator/ContractInstance"
+          "Plutus/Trace/Emulator/Extract"
           "Plutus/Trace/Emulator/System"
           "Plutus/Trace/Emulator/Types"
           "Plutus/Trace/Playground"
