+++ plutus-contract/plutus-contract.cabal
@@ -91,6 +91,7 @@ library
         Plutus.Trace.Effects.Waiting
         Plutus.Trace.Emulator
         Plutus.Trace.Emulator.ContractInstance
+        Plutus.Trace.Emulator.Extract
         Plutus.Trace.Emulator.System
         Plutus.Trace.Emulator.Types
         Plutus.Trace.Playground
@@ -123,7 +124,10 @@ library
         cryptonite >=0.25,
         data-default -any,
         deepseq -any,
+        directory -any,
+        filepath -any,
         fingertree -any,
+        flat -any,
         foldl -any,
         freer-simple -any,
         hashable -any,
